// apps/backend/src/modules/lobbies/domain/enums/room-phase.enum.ts
export enum RoomPhase {
  LOBBY = 'LOBBY',
  FORMATION_SELECTION = 'FORMATION_SELECTION',
  DRAFT = 'DRAFT',
  COACH_SELECTION = 'COACH_SELECTION',
  TEAM_REVIEW = 'TEAM_REVIEW',
  MATCHES = 'MATCHES',
  FINISHED = 'FINISHED',
}

export enum ParticipantPhaseStatus {
  IN_LOBBY = 'IN_LOBBY',
  FORMATION_SELECTED = 'FORMATION_SELECTED',
  COACH_SELECTED = 'COACH_SELECTED',
}